requirements.txt
flask
requests
python-dotenv
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu